uniempresarial/requirements.txt
django-rest-swagger==2.2.0
djangorestframework==3.9.2
Django==1.11.23
django-filter==2.0.0
djoser==1.1.5
django-cors-headers==2.5.2
pillow==6.2.1
psycopg2==2.8.1